Reviewed-on: https://qigitea.quinto.site/Sem_ecos003/H202406_Ecos_03_back/pulls/8

+++ src/main/java/com/semillero/ecosistema/repositorio/IUsuarioRepositorio.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.semillero.ecosistema.entidad.Usuario;
 
-public interface IUsuarioRepositorio extends JpaRepository<Usuario, Integer> {
+public interface IUsuarioRepositorio extends JpaRepository<Usuario, Long> {
 
 }
